% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acomodo_otros.R
\name{acomodo_otros}
\alias{acomodo_otros}
\title{Titulo Pendiente}
\usage{
acomodo_otros(datos, libro.codigos)
}
\arguments{
\item{datos}{Pendiente}

\item{libro.codigos}{Pendiente}
}
\description{
Descripci贸n pendiente
}
\examples{
# Se carga el libro de codigos
# libro.codigos<-read_xlsx("Base Abiertas Final/Relaci贸n Otros con sus columnas nuevas.xlsx")
# la estructura del archivo libro de codigos es de 5 columnas: Expresion regular otros|COD.|Desc|Columna|Count
# "Expresion regular otros" se pone la expresion regular para que el codigo encuentre la columna que contiene la codificaci贸n de otros
# "COD" es el codigo de los otros
# "Desc" es la descripci贸n del codigo
# "Columna" es la variable donde el codigo va a distribuir a los otros
# "Count" en esta columna la funcion indica cuantos valores de otros movio a la columna indicada
# otros.acomodado<-acomodo.otros(datos,libro.codigos)
}
\keyword{frecuencias}
