% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColumnNames.R
\name{cleanColumnNames}
\alias{cleanColumnNames}
\title{Función para limpiar el nombre de las columnas}
\usage{
cleanColumnNames(datos)
}
\arguments{
\item{datos}{La tabla principal i.e. "la base de datos"}
}
\description{
Esta función limpia a partir de ciertas reglas el nombre de las funciones, unicamente regresa un vector con el nombre de las columnas ya limpias
}
\examples{
#pending
}
\keyword{limpiar}
