% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaSyntaxSPSS.R
\name{creaSyntaxSPSS}
\alias{creaSyntaxSPSS}
\title{Función para crear un syntax de spss desde un datamap generado por soda}
\usage{
creaSyntaxSPSS(sPath, sFile)
}
\arguments{
\item{sPath}{El archivo descargado de soda, un "dataMap"}

\item{sFile}{El nombre del archivo a generar, recordar usar la terminación ".sps"}
}
\description{
Esta función crea un syntax de spss a partir de un DataMap de Soda
}
\keyword{spss}
