% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrind.R
\name{datagrind}
\alias{datagrind}
\title{Titulo pendiente}
\usage{
datagrind(lista, esAnidado = F, etiquetas = NULL, tablaUnica = T,
  variablesIguales = T, eVariables = NULL, colapsarPor = NULL)
}
\arguments{
\item{lista}{Descripción pendiente}

\item{esAnidado}{Descripción pendiente}

\item{etiquetas}{Descripción pendiente}

\item{tablaUnica}{Descripción pendiente}

\item{variablesIguales}{Descripción pendiente}

\item{eVariables}{Descripción pendiente}

\item{colapsarPor}{Descripción pendiente}
}
\description{
Descripción pendiente
}
\examples{
datagrind(lista, esAnidado=F, etiquetas=NULL, tablaUnica=T, variablesIguales=T, eVariables = NULL, colapsarPor = NULL)
}
\keyword{Pending}
