% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuentator_viejo.R
\name{frecuentator_viejo}
\alias{frecuentator_viejo}
\title{Función para frecuencias y pruebas de proporciones (versiòn vieja)}
\usage{
frecuentator_viejo(fTtabla, fTvariables, fTlevels = T, fbanner = NULL,
  fTponderador = NULL, fTsobreQuien = NULL, fTtotal = T,
  fTprop = F, fTusarNA = F, fTdecimales = 4)
}
\arguments{
\item{fTtabla}{La tabla principal i.e. "la base de datos"}

\item{fTvariables}{La variable o nombres de variables de quien se va a extraer informacion (i.e. frecuencias)}

\item{fTlevels}{Uso los levels de las variables en fvariables? TRUE= Uso los levels, FALSE= uso los nombres de las variables, i.e. FALSE= mi variable(s) son lógicas}

\item{fbanner}{Las variables que van por banner, en caso de que se necesite}

\item{fTponderador}{Nombre de la variable ponderador, en caso de que exista}

\item{fTsobreQuien}{Fijar un total para todos los cálculos de porcentaje}

\item{fTtotal}{Agregar una fila de total en los resultados finales?}

\item{fTprop}{Hacer prueba de proporciones? En vez de regresar la tabla de frecuencias se regresa una tabla de prueba de proporciones (igual a las de SPSS)}

\item{fTusarNA}{Frecuentator omite por default los NA de las variables que le pedimos, pero a veces se necesitan i.e. Cuando agrupo variables para Share of Mind}

\item{fTdecimales}{El redondeo de porcentaje a cuantos decimales debe ser? El default es 1}
}
\description{
Esta función obtiene frecuencias, ponderadas o no, y pruebas de proporciones para tablas tipo encuesta. La funcion necesita el paquete survey.
}
\examples{
frecuentator_viejo(fTtabla = iris,fTvariables = "Species",fTlevels = T,fbanner = "Species")
}
\keyword{frecuencias}
