% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuentatorDG.R
\name{frecuentatorDG}
\alias{frecuentatorDG}
\title{Titulo Pendiente}
\usage{
frecuentatorDG(fTtabla, fTvariables, fTlevels = T, fbanner = NULL,
  fTanidado = NULL, fTponderador = NULL, fTsobreQuien = NULL,
  fTtotal = T, fTprop = F, fTusarNA = F, fTdecimales = 4,
  fTtipo = NULL, fTunion = F, fTpctConDif = F, bases = NULL,
  etiquetas = NULL, usarbases = F, tablaUnica = T,
  variablesIguales = T, eVariables = NULL)
}
\arguments{
\item{fTtabla}{Pendiente}

\item{fTvariables}{Pendiente}

\item{fTlevels}{Pendiente}

\item{fbanner}{Pendiente}

\item{fTanidado}{Pendiente}

\item{fTponderador}{Pendiente}

\item{fTsobreQuien}{Pendiente}

\item{fTtotal}{Pendiente}

\item{fTprop}{Pendiente}

\item{fTusarNA}{Pendiente}

\item{fTdecimales}{Pendiente}

\item{fTtipo}{Pendiente}

\item{fTunion}{Pendiente}

\item{fTpctConDif}{Pendiente}

\item{bases}{Pendiente}

\item{etiquetas}{Pendiente}

\item{usarbases}{Pendiente}

\item{tablaUnica}{Pendiente}

\item{variablesIguales}{Pendiente}

\item{eVariables}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
