% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuentatorNPS.R
\name{frecuentatorNPS}
\alias{frecuentatorNPS}
\title{Titulo Pendiente}
\usage{
frecuentatorNPS(fTtabla, fTvariables, fTlevels = T, fbanner = NULL,
  fTanidado = NULL, fTponderador = NULL, fTsobreQuien = NULL,
  fTtotal = T, fTprop = F, fTusarNA = F, fTdecimales = 4,
  fTtipo = NULL, fTunion = F, fTescala = NULL, fTpctConDif = F)
}
\arguments{
\item{fTtabla}{La tabla principal i.e. "la base de datos"}

\item{fTvariables}{La variable o nombres de variables de quien se va a extraer informacion (i.e. frecuencias)}

\item{fTlevels}{Uso los levels de las variables en fvariables? TRUE= Uso los levels, FALSE= uso los nombres de las variables, i.e. FALSE= mi variable(s) son lógicas}

\item{fbanner}{Las variables que van por banner, en caso de que se necesite}

\item{fTanidado}{Pendiente}

\item{fTponderador}{Nombre de la variable ponderador, en caso de que exista}

\item{fTsobreQuien}{Fijar un total para todos los cálculos de porcentaje}

\item{fTtotal}{Agregar una fila de total en los resultados finales?}

\item{fTprop}{Hacer prueba de proporciones? En vez de regresar la tabla de frecuencias se regresa una tabla de prueba de proporciones (igual a las de SPSS)}

\item{fTusarNA}{Frecuentator omite por default los NA de las variables que le pedimos, pero a veces se necesitan i.e. Cuando agrupo variables para Share of Mind}

\item{fTdecimales}{El redondeo de porcentaje a cuantos decimales debe ser? El default es 1}

\item{fTtipo}{Pendiente}

\item{fTunion}{Pendiente}

\item{fTescala}{Pendiente}

\item{fTpctConDif}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
