% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPromediator.R
\name{fullPromediator}
\alias{fullPromediator}
\title{Titulo Pendiente}
\usage{
fullPromediator(fTtabla, fTvariable, fbanner = NULL, fTanidado = NULL,
  fTponderador = NULL, fTdecimales = 1)
}
\arguments{
\item{fTtabla}{La tabla principal i.e. "la base de datos"}

\item{fTvariable}{La variable o nombres de variables de quien se va a extraer informacion (i.e. frecuencias)}

\item{fbanner}{Las variables que van por banner, en caso de que se necesite}

\item{fTanidado}{Pendiente}

\item{fTponderador}{Nombre de la variable ponderador, en caso de que exista}

\item{fTdecimales}{El redondeo de porcentaje a cuantos decimales debe ser? El default es 1}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
