% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menciones.R
\name{menciones}
\alias{menciones}
\title{Titulo Pendiente}
\usage{
menciones(data, variable, palabrasNegras, identificador)
}
\arguments{
\item{data}{Pendiente}

\item{variable}{Pendiente}

\item{palabrasNegras}{Pendiente}

\item{identificador}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
