% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muestrator.R
\name{muestrator}
\alias{muestrator}
\title{Titulo Pendiente}
\usage{
muestrator(datos, variables, muestra, proporciones = NULL)
}
\arguments{
\item{datos}{Pendiente}

\item{variables}{Pendiente}

\item{muestra}{Pendiente}

\item{proporciones}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
muestrator(base,c("estado", "lada"),1200)
}
\keyword{frecuencias}
