% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netoSubneto.R
\name{netoSubneto}
\alias{netoSubneto}
\title{Titulo pendiente}
\usage{
netoSubneto(data, variable, filtro, banner, fTanidado = NULL,
  fTprop = FALSE, fTponderador = NULL, fTtipo = NULL, fTunion = F,
  fTpctConDif = F)
}
\arguments{
\item{data}{Descripción pendiente}

\item{variable}{Descripción pendiente}

\item{filtro}{Descripción pendiente}

\item{banner}{Descripción pendiente}

\item{fTanidado}{Descripción pendiente}

\item{fTprop}{Descripción pendiente}

\item{fTponderador}{Descripción pendiente}

\item{fTtipo}{Descripción pendiente}

\item{fTunion}{Descripción pendiente}

\item{fTpctConDif}{Descripción pendiente}
}
\description{
Descripción pendiente
}
\examples{
netoSubneto(data, variable, filtro, banner, fTanidado=NULL , fTprop=FALSE, fTponderador=NULL,  fTtipo=NULL, fTunion=F, fTpctConDif=F)
}
\keyword{Pending}
