% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizPablo.R
\name{normalizPablo}
\alias{normalizPablo}
\title{Función para hacer normalizaciones}
\usage{
normalizPablo(conteos, totales)
}
\arguments{
\item{conteos}{Pendiente}

\item{totales}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
