% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poneTruesBase.R
\name{poneTruesBase}
\alias{poneTruesBase}
\title{Titulo Pendiente}
\usage{
poneTruesBase(base, variables)
}
\arguments{
\item{x}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
