% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poneTruesVar.R
\name{poneTruesVar}
\alias{poneTruesVar}
\title{Titulo Pendiente}
\usage{
poneTruesVar(x)
}
\arguments{
\item{x}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
pendiente
}
\keyword{frecuencias}
