% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promedietor.R
\name{promedietor}
\alias{promedietor}
\title{Función para basada en frecuentator para sacar promedios}
\usage{
promedietor(fTtabla, fTvariable, fbanner = NULL, fTponderador = NULL,
  fTdecimales = 1)
}
\arguments{
\item{fTtabla}{La tabla principal i.e. "la base de datos"}

\item{fTvariable}{La variable de quien se va a extraer informacion (i.e. promedio)}

\item{fbanner}{Las variables que van por banner, en caso de que se necesite}

\item{fTponderador}{Nombre de la variable ponderador, en caso de que exista}

\item{fTdecimales}{El redondeo de porcentaje a cuantos decimales debe ser? El default es 1}
}
\description{
Esta función obtiene promedios, ponderados o no
}
\examples{
turboP3 <- lapply(misP3, function(x){promedietor(fTtabla = datos,fTvariable = x,fTdecimales = 2)})
}
\keyword{promedio}
