% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeador.R
\name{rangeador}
\alias{rangeador}
\title{Agrupa resultados "numericos" de un data.frame de frecuentator}
\usage{
rangeador(rtarget, rrangos)
}
\arguments{
\item{rtarget}{El data frame generado por frecuentator}

\item{rrangos}{Un data frame que tiene definidos los rangos en que se va a agrupar cada fila de rtarget}
}
\description{
Dado un data.frame generado por frecuentator, tal que la respuesta sea númerica, es posible agruparla en rangos con esta función
}
\examples{
drangos1 <- data.frame(de=c(0,1,5,10,15,20),a=c(1,5,10,15,20,0))
drangos2 <- data.frame(de=c(0,1000,2000,3000,4000,5000),a=c(1000,2000,3000,4000,5000,0))
drangos3 <- data.frame(de=c(0,80000,160000,240000,400000),a=c(80000,160000,240000,400000,0))
D8_Cuál_es_el_plazo_de_su_seguro_de_vida_TOTAL=rangeador(rtarget = brostatistics::frecuentator(fTtabla = datosb,fTvariables = nombresR(datosb,"D8_.*?r1_\\\\d$"),fTlevels = T,fbanner = bandera,fTusarNA = T),rrangos = drangos1)
}
\keyword{agrupar}
