% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_tablas.R
\name{revision_tablas}
\alias{revision_tablas}
\title{Tablas de}
\usage{
revision_tablas(data, regex)
}
\arguments{
\item{data}{Base de datos}

\item{regex}{Regex que busca en nombresR}
}
\description{
Tablas de
}
