% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablagrid.R
\name{tablagrid}
\alias{tablagrid}
\title{Creación de tablas ponderadas o no}
\usage{
tablagrid(base, filas = NA, columnas = NA, esquema = NA,
  tipo = "FP", ordenar = FALSE, ponderador = NA, prop = "NIVELES",
  filtro.filas = c("", " ", "-"), filtro.columnas = NA,
  base.natural = FALSE, usarNA = FALSE, remover.duplicados = TRUE)
}
\arguments{
\item{base}{el data frame que contiene la base.}

\item{filas}{La variable o variables que se mostraran en las filas, para anidar variables se utilza ":"\code{filas = "Género:Edad"} , puede ser NA cuando se incluye el esquema.}

\item{columnas}{La variable o variables que se mostraran en las columnas, para anidar variables se utilza ":"\code{columnas = "Género:Edad"} , puede ser NA cuando se incluye el esquema.}

\item{esquema}{Las especificaciones para construir tablas grid.}

\item{tipo}{El tipo de tabla resultante, puede generar tablas de frecuencias \code{tipo = "F"}, porcentajes \code{tipo = "P"},
diferencias \code{tipo = "D"}, o tablas combinadas frecuencias y porcentajes \code{tipo =  "FP"}, porcentajes y diferencias \code{tipo = "PD"}.}

\item{ordenar}{Se puede ordenar la tabla de forma descentente \code{ordenar = "D"} o ascendente \code{ordenar = "A"},
para ordenar requiere que exista una variable de total en las columnas.}

\item{ponderador}{Nombre de la variable ponderador, en caso de que exista.}

\item{prop}{en este parametro se especifica si la prueba de proporciones se realizara sobre los niveles internos de cada variable
\code{diferencias = NIVELES} o sobre todas las variables ingresadas (ignorando el total en caso de que exista) \code{diferencias = COLUMNAS}.}

\item{filtro.filas}{niveles para remover de las filas ingresadas, de forma predeterminada \code{filtro.filas = c(""," ","-")}.}

\item{filtro.columnas}{niveles para remover de las columnas ingresadas.}

\item{base.natural}{usar \code{base.natural = TRUE} en caso de que se desee tener la base natural en la tabla resultante.}

\item{usarNA}{si se van a incluir los valores \code{NA} en la tabla.}

\item{remover.duplicados}{si se van a remover los valores duplicados en las filas.}
}
\value{
devuelve una tabla de frecuencias de acuerdo a lo especificado en el parametro "tipo".
}
\description{
Creación de tablas ponderadas o no
}
\examples{
tablagrid(iris,filas = "Species",columnas ="Species",ordenar = "D")
#Para usar el esquema suponiendo que se tiene lo siguiente:
# filas<-c("P9","P14","P19","P24","P30")
# columnas<-c("Total","P8","P13","P18","P23","P29")
# esquema<-c(paste0("Total|",paste0(filas, collapse=",")),"P8|P9","P13|P14","P18|P19","P23|P24","P29|P30")

}
