% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnionMenciones.R
\name{UnionMenciones}
\alias{UnionMenciones}
\title{Titulo Pendiente}
\usage{
UnionMenciones(base, variables, identificador, palabrasNegras)
}
\arguments{
\item{base}{Pendiente}

\item{variables}{Pendiente}

\item{identificador}{Pendiente}

\item{palabrasNegras}{Pendiente}
}
\description{
Descripción pendiente
}
\examples{
iris$ID<-1:nrow(iris)
iris$Species2<-iris$Species
UnionMenciones(base=iris,variables = c('Species'),identificador = 'ID', palabrasNegras = c(''))
}
\keyword{frecuencias}
