% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poneTruesBase.R
\name{poneTruesBase}
\alias{poneTruesBase}
\title{Convertir columna en logica}
\usage{
poneTruesBase(base, variables, extra = c())
}
\arguments{
\item{base}{La base a considerar}

\item{variables}{Las variables a convertir, meterla como vector de strings}

\item{extra}{Algunos caracteres/texto extra para poner FALSE en estos casos}
}
\value{
Data.frame
}
\description{
Convertir columna en logica
}
